requirements.txt
requests==2.32.3
PyPDF2==3.0.1
beautifulsoup4==4.12.3
python-dotenv==1.0.1
google-generativeai==0.8.3
ruff==0.11.2
black==25.1.0
mypy==1.15.0
Flask==3.1.0
markdown==3.7
mkdocs==1.6.1
mkdocstrings-python==1.16.8
mkdocs-material==9.6.10
Flask-SocketIO==5.5.1
chromadb==1.0.0
langchain==0.3.23